Safari-611.1.10/LayoutTests/svg/animations/svgnumberlist-animation-1-expected.txt

SVG 1.1 dynamic animation tests

ABCD
Test SVGNumberList animation on 'rotate' attribute of text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.rotate.animVal.numberOfItems is 1
PASS text.rotate.animVal.getItem(0).value is 15
PASS text.rotate.baseVal.numberOfItems is 1
PASS text.rotate.baseVal.getItem(0).value is 15
PASS text.rotate.animVal.numberOfItems is 4
PASS text.rotate.animVal.getItem(0).value is 0
PASS text.rotate.animVal.getItem(1).value is 0
PASS text.rotate.animVal.getItem(2).value is 0
PASS text.rotate.animVal.getItem(3).value is 0
PASS text.rotate.baseVal.numberOfItems is 1
PASS text.rotate.baseVal.getItem(0).value is 15
PASS text.rotate.animVal.numberOfItems is 4
PASS text.rotate.animVal.getItem(0).value is 22.5
PASS text.rotate.animVal.getItem(1).value is 45
PASS text.rotate.animVal.getItem(2).value is 67.5
PASS text.rotate.animVal.getItem(3).value is 90
PASS text.rotate.baseVal.numberOfItems is 1
PASS text.rotate.baseVal.getItem(0).value is 15
PASS text.rotate.animVal.numberOfItems is 4
PASS text.rotate.animVal.getItem(0).value is 45
PASS text.rotate.animVal.getItem(1).value is 90
PASS text.rotate.animVal.getItem(2).value is 135
PASS text.rotate.animVal.getItem(3).value is 180
PASS text.rotate.baseVal.numberOfItems is 1
PASS text.rotate.baseVal.getItem(0).value is 15
PASS text.rotate.animVal.numberOfItems is 1
PASS text.rotate.animVal.getItem(0).value is 15
PASS text.rotate.baseVal.numberOfItems is 1
PASS text.rotate.baseVal.getItem(0).value is 15
PASS successfullyParsed is true

TEST COMPLETE
